requirements_rag.txt
flask
gunicorn
flask-login
groq
python-dotenv
psycopg[binary]
pypdf
pytesseract
pdf2image
qdrant-client
sentence-transformers
supabase
requests
